+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Removed expired flow limitation challenges
 - **Admin Changes**
   - Improved Downloads by Quarter report loading speed
+- **API Changes**
+  - Dataset versions now have their release date set on creation
 - **Data Request Changes**
   - Data requests now use
     [SignaturePad](https://github.com/szimek/signature_pad)